+++ modulo4/projeto-labex/src/routes/Router.jsx
@@ -2,7 +2,7 @@ import { Route, Switch } from "react-router-dom";
 import { BrowserRouter } from "react-router-dom/cjs/react-router-dom.min";
 import AdminPage from "../pages/AdminPage";
 import HomePage from "../pages/HomePage";
-import TravelsPage from "../pages/TravelsPage";
+import TripsPage from "../pages/TripsPage";
 
 const Router = () => {
   return (
@@ -12,8 +12,8 @@ const Router = () => {
           <HomePage />
         </Route>
 
-        <Route exact path={"/travels"}>
-          <TravelsPage />
+        <Route exact path={"/trips"}>
+          <TripsPage />
         </Route>
 
         <Route exact path={"/login"}>
